Plug soundness hold around mem::forgetting MoveRefs
The Pin drop invariant requires that pinned data be properly destroyed
before its memory be re-used. This is a problem: MoveRef is responsible
for running the destructor of storage it generally does not own. This
means we must leak the storage of a forgotten MoveRef. This commit
introduces explicit library-tracked drop flags to manage this corner
case.

In the case of stack-pinned MoveRefs via `slot`, we can get into
situations (such as if a constructor panics) where we cannot properly
destroy the pointee. For now, we double-panic to force an abort, but we
may be able to handle this better in the future.