Remove unsound `PinExt::as_move` function
This function cannot possibly be sound: it wants to move out of a
MoveRef it owned and return a MoveRef to its pointer, but if this were
possible, DerefMove would be a single function, not two.

For now, we won't reproduce this functionality, but it may be re-added
as a macro later on, e.g. `&move pin *x` or similar.