Add drop flags to `MoveRef`
This change adds a "drop flag" field to `MoveRef`, allowing it to behave
more like a reference to a stack slot. In particular this ensures that
stack-allocated values that have gone on to be pinned cannot be
forgotten later. Although this is ostensibly a small pessimization, it
should be minor compared to the cost of something being destroyed
through a `MoveRef`, which will frequently be a possibly-complex type.