Implement DerefMove for various types.
The previous change ceased to take any Into<Pin<P>> for `mov`.
We therefore need to explicitly implement `DerefMove` for any type
that may be passed into `mov`.

It would be ideal to add a blanket implementation of DerefMove
for any type which is Into<Pin<P>>, something like this...

  unsafe impl<P, T> DerefMove for T
  where T: Into<Pin<P>>,
    P: Deref<Target=T>,
    P: DerefMove {
      // ...
  }

but the type parameters are insufficiently constrained so this
seems impossible.

A second best would be to implement DerefMove for any T: Unpin
rather than &T, &mut T. This too is impossible because we get
conflicting implementations of DerefMove.

Hence this commit settles on implementing DerefMove for:

  MoveRef<T> (was previously present)
  Pin<&T> where T: Unpin
  Pin<&mut T> where T: Unpin
  Pin<MoveRef<T>>
  Box<T> (was previously present)
  Pin<Box<T>>

This should give roughly the same flexibility for callers of
`mov`.