Remove unsound/problematic DerefMove for Pin types
The `Pin<Box<T>>: DerefMove` impl is unsound because it allows you to
obtain a `MoveRef<T>` from `Pin<Box<T>>` even if `T: !Unpin`.

Once you have the `MoveRef<T>` you can then (without any `unsafe`)
either get a `T` (via `MoveRef::into_inner`) or `&mut T`, which then
lets you trivially code up the problematic `std::mem::swap` example
motivating the `Pin` type in the Rust docs.

The `Pin<&'a T>, Pin<&mut 'a T>: DerefMove` impl's are problematic for a
different reason.

If we examine the definition of `deref_move` for these impls, we see:

  fn deref_move<'frame>(
    self,
    storage: DroppingSlot<'frame, Self::Storage>,
  ) -> MoveRef<'frame, Self::Target>
  where
    Self: 'frame,
  {
    Pin::into_inner(
      MoveRef::into_pin(
        DerefMove::deref_move(self, storage)
      )
    )
  }

This definition type checks, but in practice it will fail by blowing the
stack because it is infinitely recursive.

The inner-most call to `DerefMove::deref_move(self, storage)` will
resolve to the exact same method currently being defined. In other
words, the call-graph will look like:

  Pin::into_inner(
    MoveRef::into_pin(
      Pin::into_inner(
        MoveRef::into_pin(
          ...
        )
      )
    )
  )

But not actually do anything productive before crashing.

This issue is partially hidden by redundancy in the bounds, which are
written as this:

  Pin<&'a T>: DerefMove + DerefMut,
  Pin<&'a T>: Deref<Target = T>,
  T: Unpin,

The `Pin<&'a T>: DerefMove` bound is redundant (since the impl being
defined is itself `Pin<&'a T>: DerefMove`).

Indeed, if you remove the redundant `Pin<&'a T>: DerefMove`, the
compiler immediately gives an error about the unbounded recursion:

  function cannot return without recursing
  a `loop` may express intention better if this is on purpose
  `#[deny(unconditional_recursion)]` implied by `#[deny(warnings)]`

But also it would make more sense to express the bounds like this:

  Self: DerefMut,
  Self: Deref<Target = T>,
  T: Unpin,

I assume the issue with these `Pin<&'a T>: DerefMove` impls hasn't been
noticed because they are not used anywhere in the current code base.

In fact, it's not even *possible* to use them in their current form, due
to the redundant bound, which causes the compiler to fail when it tries
to resolve the `DerefMove` impl for `Pin<&T>`.

As a demonstration of this, consider the following tests:

  #[test]
  fn bad_pin_deref_move() {
    let val = ();
    let val = unsafe { core::pin::Pin::new_unchecked(&val) };
    crate::slot!(
      #[dropping]
      storage
    );
    val.deref_move(storage);
  }

  #[test]
  fn bad_pin_mut_deref_move() {
    let mut val = ();
    let val = unsafe { core::pin::Pin::new_unchecked(&mut val) };
    crate::slot!(
      #[dropping]
      storage
    );
    val.deref_move(storage);
  }

Prior to these commits, these examples don't even compile, but they
should. Instead, the compiler complains with the following errors:

  error[E0599]: the method `deref_move` exists for struct `Pin<&()>`,
  but its trait bounds were not satisfied
     --> src/move_ref.rs:673:9
      |
  673 |     val.deref_move(storage);
      |         ^^^^^^^^^^ method cannot be called on `Pin<&()>` due to
      unsatisfied trait bounds

  pub struct Pin<P> {
      | -----------------
      | |
      | doesn't satisfy `std::pin::Pin<&()>: DerefMut`
      | doesn't satisfy `std::pin::Pin<&()>: move_ref::DerefMove`
      |
  note: the following trait bounds were not satisfied:
        `std::pin::Pin<&()>: DerefMut`
        `std::pin::Pin<&()>: move_ref::DerefMove`
     --> src/move_ref.rs:326:15
      |
  324 | unsafe impl<'a, T> DerefMove for Pin<&'a T>
      |                    ---------     ----------
  325 | where
  326 |   Pin<&'a T>: DerefMove + DerefMut,
      |               ^^^^^^^^^   ^^^^^^^^ unsatisfied trait bound
      introduced here
      |               |
      |               unsatisfied trait bound introduced here

But this is clearly wrong because `std::pin::Pin<&()>` *does* have
impls for `DerefMut` and `move_ref::DerefMove`.

The problem is the redundant bound I pointed out earlier. We can change
the impl definition to this:

  unsafe impl<'a, T> DerefMove for Pin<&'a T>
  where
    Self: Deref<Target = T>,
    T: Unpin,
  {
    #[allow(unconditional_recursion)]
    #[inline]
    fn deref_move<'frame>(
      self,
      storage: DroppingSlot<'frame, Self::Storage>,
    ) -> MoveRef<'frame, Self::Target>
    where
      Self: 'frame,
    {
      Pin::into_inner(
        MoveRef::into_pin(
          DerefMove::deref_move(self, storage)
        )
      )
    }
  }

Now the test compiles and runs, but as expected, overflows the stack:

  thread 'move_ref::test::bad_pin_deref_move' has overflowed its stack
  fatal runtime error: stack overflow