Restore original `Pin<P>: DerefMove` impl
This commit restores the original `Pin<P>: DerefMove` impl prior to the
commit 1b570c5.

This original impl does not have the soundness hole described in the
previous commit, since it is a blanket impl for all `Pin<P>` and
requires that `P::Target: Unpin`.

The changes to `DerefMove` in 1b570c5 were motivated by ergonomics
around using `cxx::UniquePtr` with `mov` in this crate, but we enable
similar functionality in a different (and hopefully sound) way in
subsequent commits by modifying `mov` to take `ptr: impl PinExt<P>`
and then defining `PinExt` impls for `MoveRef` and `UniquePtr`.