Rename PinExt to AsMove, change interface to mov, fixes for UniquePtr
This commit renames the `PinExt` trait to `AsMove` and changes the
interface for `mov` to `ptr: impl AsMove`.

The parameter `P` for `PinExt<P>` has been removed and the bound
`P: DerefMove` relaxed to where now it is only required that
`AsMove: Deref`.

`DerefMove` is now modified so that `AsMove` is a super trait, i.e.,
`DerefMove: AsMove`. Relatedly, the associated `DerefMove::Storage`
type has been moved to `AsMove::Storage`.

This refactoring is to facilitate changes to the `DerefMove` impl for
`UniquePtr<T>`, which needs `T: Unpin` similarly to how the `DerefMove`
impl for `Pin<P>` needs `P::Target: Unpin`, otherwise the same soundness
issue that was present with `Pin<Box<T>>` is also present for
`UniquePtr<T>`.

This means that when `T: !Unpin`, given `ptr: UniquePtr<T>`, a
`moveit!(let x = &move *ptr;)` is rejected now (since `DerefMove` is
not defined), but (due to the relaxed bound), it is still possible to
have `moveit!(moveit::new::mov(ptr))` (since `AsMove` is defined).