Rename PinExt to AsMove, change interface to mov, fixes for UniquePtr
This commit renames the `PinExt` trait to `AsMove` and changes the
interface to `mov` to `ptr: impl AsMove<P>`.

The bound on `P` for `AsMove<P>` is also relaxed from `DerefMove` to
just `Deref`.

This is to facilitate changes to the `DerefMove` impl for
`UniquePtr<T>`, which needs `T: Unpin` similarly to how the `DerefMove`
impl for `Pin<P>` needs `P::Target: Unpin`, otherwise the same soundness
issue that was present with `Pin<Box<T>>` is also present for
`UniquePtr<T>`.

This means that when `T: !Unpin`, given `ptr: UniquePtr<T>`, a
`moveit!(let x = &move *ptr;)` is rejected now, but
(due to the relaxed bound), it is still possible to have
`moveit!(moveit::new::mov(ptr))`.